requirements.txt
asgiref==3.7.2
Django==4.1.13
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.4.2
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.2.4
typing_extensions==4.8.0
